[bridge-2452] configure https redirect at the ALB
Use the AWS application load balancer to handle https redirect.
This is not a great solution because we don't have a way to
automatically inject the LoadBalancerArn into the template.
AWS Beanstalk does not provide a reference or option to configure
the ALB that it provisions.  This implementation requires us to
manually look up the ALB Arn post beanstalk deployment and pass
the Arn to the template.